+++ app/src/lib/components/settings/Sidebar.svelte
